Field Lists: Add field-lists option.
Adds field-lists cli option that adds field_list field to each summary
table.

This also modifies two summary tables
contract_implementation_transactions_summary and planning_summary.
This is due to them not already having a field that contains a JSONB
of the data, which is inconsistant with all the other tables.
planning_summary is now a view, in order not to copy the whole of the
planning data.

A global SUMMARY_TABLES list has been made in manage.py as there is no
simple way to gather these definitions from the schema itself and there
is no easy way to gather the fields contained from a naming pattern.
This list has also been used in test_add so there is no need to list out
each summary_table manually there.